+++ packages/pino-adapter/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @korix/pino-adapter
 
+## 0.1.0-alpha.3
+
+### Patch Changes
+
+- Refactor pino-adapter to move logger factory logic out of index.ts
+
 ## 0.1.0-alpha.2
 
 ### Patch Changes